+++ shards/rust/src/core.rs
@@ -29,7 +29,7 @@ use std::os::raw::c_char;
 
 const ABI_VERSION: u32 = 0x20200101;
 
-pub const MIN_STACK_SIZE: usize = crate::shardsc::SH_MIN_STACK_SIZE;
+pub const MIN_STACK_SIZE: usize = crate::shardsc::SH_MIN_STACK_SIZE as usize;
 
 pub static mut Core: *mut SHCore = core::ptr::null_mut();
 pub static mut ScriptEnvCreate: Option<
